+++ internal/xray/tun_other.go
@@ -5,3 +5,7 @@ package xray
 func prepareTUN() error {
 	return nil
 }
+
+func missingTUNDLL() bool {
+	return false
+}